move Image change build triggers
